requirement.txt
asgiref==3.3.4
autopep8==1.5.7
certifi==2021.5.30
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
coverage==5.5
Django==3.2.4
django-cors-headers==3.7.0
django-filter==2.4.0
django-rest-swagger==2.2.0
djangorestframework==3.12.4
drf-yasg==1.20.0
idna==2.10
importlib-metadata==4.5.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
Markdown==3.3.4
MarkupSafe==2.0.1
openapi-codec==1.3.2
packaging==20.9
pkg-resources==0.0.0
pycodestyle==2.7.0
pyparsing==2.4.7
python-dotenv==0.17.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
ruamel.yaml==0.17.9
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.5
zipp==3.4.1
